refactor: merge component_proxy + component_resolver into builder/_component.py
Both modules were used only by builder/_dispatch_mixin. Merge into
a single internal module in the builder package.
