+++ src/genro_builders/__init__.py
@@ -39,10 +39,9 @@ from genro_builders.binding import (
     scan_for_pointers,
 )
 from genro_builders.builder import BagBuilderBase
+from genro_builders.builder._component import ComponentProxy, ComponentResolver
 from genro_builders.builder_bag import BuilderBag, BuilderBagNode, Component
 from genro_builders.compiler import BagCompilerBase, compiler
-from genro_builders.component_proxy import ComponentProxy
-from genro_builders.component_resolver import ComponentResolver
 from genro_builders.contrib.yaml import YamlRendererBase
 from genro_builders.manager import BuilderManager, ReactiveManager
 from genro_builders.renderer import BagRendererBase, RenderNode, renderer
