+++ src/genro_builders/builder/_component.py
@@ -1,23 +1,14 @@
 # Copyright 2025 Softwell S.r.l. - SPDX-License-Identifier: Apache-2.0
-"""ComponentResolver - lazy expansion of components via BagResolver.
+"""Component infrastructure — proxy and lazy resolver.
 
-A ComponentResolver is attached to each component node at creation time.
-When the node is accessed with static=False (via walk, expand, or compile),
-the resolver executes the component handler body and returns the populated Bag.
+ComponentProxy: transparent proxy returned by component calls.
+    Wraps the parent Bag (root) and optional named slot Bags.
+    Attribute access delegates to root, except for slot names.
 
-Supports component inheritance via ``based_on``: a derived component receives
-the Bag already populated by its parent component, then applies its own
-modifications.
-
-Example:
-    >>> class MyBuilder(BagBuilderBase):
-    ...     @component()
-    ...     def base_form(self, comp, **kwargs):
-    ...         comp.input(name='field1')
-    ...
-    ...     @component(based_on='base_form')
-    ...     def extended_form(self, comp, **kwargs):
-    ...         comp.input(name='field2')  # adds to base_form content
+ComponentResolver: lazy expansion of components via BagResolver.
+    Attached to component nodes at creation time. Executes the
+    handler body on first non-static access (walk, expand, compile).
+    Supports inheritance via ``based_on``.
 """
 from __future__ import annotations
 
@@ -25,6 +16,70 @@ from typing import Any
 
 from genro_bag import Bag
 from genro_bag.resolver import BagResolver
+
+# ---------------------------------------------------------------------------
+# ComponentProxy
+# ---------------------------------------------------------------------------
+
+
+class ComponentProxy:
+    """Transparent proxy for component calls with optional named slots.
+
+    Delegates attribute access to root Bag. Slot names are intercepted
+    and return the corresponding slot Bag instead.
+    """
+
+    def __init__(self, root: Any, slots: dict[str, Any] | None = None) -> None:
+        object.__setattr__(self, "_root", root)
+        object.__setattr__(self, "_slots", slots or {})
+
+    def __getattr__(self, name: str) -> Any:
+        slots = object.__getattribute__(self, "_slots")
+        if name in slots:
+            return slots[name]
+        root = object.__getattribute__(self, "_root")
+        return getattr(root, name)
+
+    def __setattr__(self, name: str, value: Any) -> None:
+        root = object.__getattribute__(self, "_root")
+        setattr(root, name, value)
+
+    def __getitem__(self, key: Any) -> Any:
+        root = object.__getattribute__(self, "_root")
+        return root[key]
+
+    def __setitem__(self, key: Any, value: Any) -> None:
+        root = object.__getattribute__(self, "_root")
+        root[key] = value
+
+    def __len__(self) -> int:
+        root = object.__getattribute__(self, "_root")
+        return len(root)
+
+    def __iter__(self) -> Any:
+        root = object.__getattribute__(self, "_root")
+        return iter(root)
+
+    def __dir__(self) -> list[str]:
+        """Return slot names plus root's dir for autocompletion."""
+        slots = object.__getattribute__(self, "_slots")
+        root = object.__getattribute__(self, "_root")
+        base = set(dir(root))
+        base.update(slots.keys())
+        return sorted(base)
+
+    def __repr__(self) -> str:
+        slots = object.__getattribute__(self, "_slots")
+        root = object.__getattribute__(self, "_root")
+        if slots:
+            slot_names = ", ".join(sorted(slots))
+            return f"<ComponentProxy slots=[{slot_names}] root={root!r}>"
+        return f"<ComponentProxy root={root!r}>"
+
+
+# ---------------------------------------------------------------------------
+# ComponentResolver
+# ---------------------------------------------------------------------------
 
 
 class ComponentResolver(BagResolver):
@@ -35,12 +90,6 @@ class ComponentResolver(BagResolver):
 
     With ``based_on``, resolves the parent component first, then passes
     the already-populated Bag to the current handler for modification.
-
-    Attributes (via class_kwargs):
-        handler: The component handler callable (bound method on builder).
-        builder_class: Builder class for creating the component's internal Bag.
-        based_on: Name of the parent component (for inheritance chain).
-        builder: Builder instance (needed to resolve based_on via schema).
     """
 
     class_kwargs: dict[str, Any] = {
@@ -59,39 +108,28 @@ class ComponentResolver(BagResolver):
     }
 
     def load(self) -> Bag:
-        """Execute the component handler and return the populated Bag.
-
-        If based_on is set, resolves the parent component first (recursively),
-        then passes the result to the current handler for modification.
-
-        If the handler returns a dict (slot mapping), mounts the content
-        from each slot Bag into the corresponding destination Bag.
-        """
+        """Execute the component handler and return the populated Bag."""
         handler = self._kw["handler"]
         builder_class = self._kw["builder_class"]
         based_on = self._kw["based_on"]
         builder_instance = self._kw["builder"]
         slots = self._kw.get("slots") or {}
 
-        # Collect kwargs from parent node attributes
         kwargs = dict(self._parent_node.attr) if self._parent_node else {}
 
         if based_on:
-            # Resolve parent component first
             comp_bag = self._resolve_parent(based_on, builder_instance, kwargs)
         else:
-            from .builder_bag import Component
+            from ..builder_bag import Component
 
             comp_bag = Component(builder=builder_class)
             comp_bag._skip_parent_validation = True
 
-        # Call the current handler
         result = None
         if handler:
             result = handler(comp_bag, **kwargs)
 
         # Mount slot content into destination Bags
-        # Use Bag.set_item directly to bypass builder interception
         if slots and isinstance(result, dict):
             from genro_bag import BagNode
 
@@ -100,10 +138,9 @@ class ComponentResolver(BagResolver):
                 if source_bag is None or len(source_bag) == 0:
                     continue
 
-                # dest can be a BagNode (from comp.container()) or a Bag
                 if isinstance(dest, BagNode):
                     if not isinstance(dest.value, Bag):
-                        from .builder_bag import BuilderBag
+                        from ..builder_bag import BuilderBag
                         dest.value = BuilderBag(builder=builder_class)
                     dest_bag = dest.value
                 else:
@@ -124,7 +161,7 @@ class ComponentResolver(BagResolver):
         self, based_on: str, builder_instance: Any, kwargs: dict[str, Any],
     ) -> Bag:
         """Resolve the parent component, handling recursive based_on chains."""
-        from .builder_bag import Component
+        from ..builder_bag import Component
 
         parent_info = builder_instance._get_schema_info(based_on)
         parent_handler_name = parent_info.get("handler_name")
@@ -133,7 +170,6 @@ class ComponentResolver(BagResolver):
         parent_based_on = parent_info.get("based_on")
 
         if parent_based_on:
-            # Recursive: parent is also based on another component
             comp_bag = self._resolve_parent(parent_based_on, builder_instance, kwargs)
         else:
             comp_bag = Component(builder=parent_builder_class)
